Added the external connections to the location block
